Backend/requirements.txt
python-dotenv
psycopg2
uvicorn
fastapi
pandas
fsspec
s3fs